+++ AmazonChimeSDK/AmazonChimeSDKTests/internal/audio/DefaultAudioClientObserverTests.swift
@@ -196,6 +196,69 @@ class DefaultAudioClientObserverTests: XCTestCase {
 
         wait(for: [expect], timeout: defaultTimeout)
     }
+    
+    func testAudioClientStateChanged_ConnectionFailedFromReconnecting() {
+        given(audioClientMock.stopSession()).willReturn(0)
+        DefaultAudioClientController.state = .started
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_RECONNECTING,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_DISCONNECTED_ABNORMAL,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.audioServerHungup.rawValue))
+        let expect = eventually {
+            verify(mockAudioVideoObserver.audioSessionDidCancelReconnect()).wasCalled()
+            verify(eventAnalyticsControllerMock.publishEvent(name: .meetingFailed, attributes: any())).wasCalled()
+            verify(meetingStatsCollectorMock.resetMeetingStats()).wasCalled()
+        }
+
+        wait(for: [expect], timeout: defaultTimeout)
+    }
+    
+    func testAudioClientStateChanged_FinishDisconnectingFromConnecting() {
+        given(audioClientMock.stopSession()).willReturn(0)
+        DefaultAudioClientController.state = .started
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_CONNECTING,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_SERVER_HUNGUP,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.audioServerHungup.rawValue))
+        let expect = eventually {
+            verify(eventAnalyticsControllerMock.publishEvent(name: .meetingEnded, attributes: any())).wasCalled()
+            verify(mockAudioVideoObserver.audioSessionDidStopWithStatus(sessionStatus: any())).wasCalled()
+        }
+
+        wait(for: [expect], timeout: defaultTimeout)
+    }
+
+    func testAudioClientStateChanged_FinishDisconnectingFromConnected() {
+        given(audioClientMock.stopSession()).willReturn(0)
+        DefaultAudioClientController.state = .started
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_CONNECTED,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_SERVER_HUNGUP,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.audioServerHungup.rawValue))
+        let expect = eventually {
+            verify(mockAudioVideoObserver.audioSessionDidStopWithStatus(sessionStatus: any())).wasCalled()
+            verify(eventAnalyticsControllerMock.publishEvent(name: .meetingEnded, attributes: any())).wasCalled()
+            verify(meetingStatsCollectorMock.resetMeetingStats()).wasCalled()
+        }
+
+        wait(for: [expect], timeout: defaultTimeout)
+    }
+    
+    func testAudioClientStateChanged_FinishDisconnectingFromReconnecting() {
+        given(audioClientMock.stopSession()).willReturn(0)
+        DefaultAudioClientController.state = .started
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_RECONNECTING,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.ok.rawValue))
+        defaultAudioClientObserver.audioClientStateChanged(AUDIO_CLIENT_STATE_SERVER_HUNGUP,
+                                                           status: audio_client_status_t.init(MeetingSessionStatusCode.audioServerHungup.rawValue))
+        let expect = eventually {
+            verify(mockAudioVideoObserver.audioSessionDidCancelReconnect()).wasCalled()
+            verify(eventAnalyticsControllerMock.publishEvent(name: .meetingEnded, attributes: any())).wasCalled()
+            verify(meetingStatsCollectorMock.resetMeetingStats()).wasCalled()
+        }
+
+        wait(for: [expect], timeout: defaultTimeout)
+    }
 
     func testAudioClientStateChanged_NotifiesOfInputDeviceFailure() {
         given(audioClientMock.stopSession()).willReturn(0)